+++ Sources/SuperwallKit/Paywall/Presentation/PaywallInfo.swift
@@ -232,7 +232,7 @@ public final class PaywallInfo: NSObject {
     forProduct product: StoreProduct? = nil,
     otherParams: [String: Any]? = nil
   ) async -> [String: Any] {
-    var output = customParams()
+    var output = audienceFilterParams()
 
     output += [
       "paywalljs_version": paywalljsVersion as Any,
@@ -256,7 +256,8 @@ public final class PaywallInfo: NSObject {
       "experiment_id": experiment?.id as Any,
       "variant_id": experiment?.variant.id as Any,
       "cache_key": cacheKey,
-      "build_id": buildId
+      "build_id": buildId,
+      "close_reason": closeReason.description
     ]
 
     var loadingVars: [String: Any] = [:]
@@ -294,8 +295,8 @@ public final class PaywallInfo: NSObject {
     return output
   }
 
-  /// Parameters that can be used in rules.
-  func customParams() -> [String: Any] {
+  /// Parameters that can be used in audience filters.
+  func audienceFilterParams() -> [String: Any] {
     var output: [String: Any] = [
       "paywall_id": databaseId,
       "paywall_name": name,
@@ -330,7 +331,6 @@ public final class PaywallInfo: NSObject {
 // MARK: - Stubbable
 extension PaywallInfo: Stubbable {
   static func stub() -> PaywallInfo {
-    let dependencyContainer = DependencyContainer()
     return PaywallInfo(
       databaseId: "abc",
       identifier: "1",